+++ packages/core/src/m3/serializer.ts
@@ -1,5 +1,5 @@
 import { LionWebJsonChunk } from "@lionweb/json"
-import { nodeSerializer } from "../serializer.js"
+import { serializerWith } from "../serializer.js"
 import { lioncoreReader } from "./facade.js"
 import { Language } from "./types.js"
 
@@ -9,5 +9,5 @@ import { Language } from "./types.js"
  * into the LionWeb serialization JSON format.
  */
 export const serializeLanguages = (...languages: Language[]): LionWebJsonChunk =>
-    nodeSerializer(lioncoreReader)(languages)
+    serializerWith({ reader: lioncoreReader })(languages)