+++ packages/core/src/serializer.ts
@@ -19,6 +19,12 @@ import {
 
 
 /**
+ * Type definition for functions that serializes nodes as a {@link LionWebJsonChunk serialization chunk}.
+ */
+export type Serializer<NT extends Node> = (nodes: NT[]) => LionWebJsonChunk
+
+
+/**
  * Interface for objects that expose a method to serialize a property's value.
  */
 export interface PropertyValueSerializer {
@@ -50,7 +56,7 @@ export type SerializationOptions = Partial<{
 
     /**
      * A {@link PropertyValueSerializer} implementation.
-     * Default = DefaultPropertyValueSerializer.
+     * Default = {@link builtinPropertyValueSerializer}.
      */
     propertyValueSerializer: PropertyValueSerializer
 
@@ -62,14 +68,41 @@ export type SerializationOptions = Partial<{
 }>
 
 
+/**
+ * Type for objects to configure {@link Serializer node serializers} with.
+ * The `reader` property is mandatory,
+ * and the `serializeEmptyFeatures`, `propertyValueSerializer`,
+ * and `primitiveTypeSerializer` (which is a legacy alias for `propertyValueSerializer`)
+ * properties are optional, with defined defaults.
+ */
+export type SerializerConfiguration<NT extends Node> = {
+    /**
+     * An interface with functions to “read” – i.e., introspect – nodes.
+     */
+    reader: Reader<NT>
+} & SerializationOptions
+
 
 /**
- * @return a function that serializes the {@link Node nodes} passed to it.
+ * @return a {@link Serializer} function that serializes the {@link Node nodes} passed to it,
+ * configured through a `reader` {@link Reader} instance,
+ * and (optionally) a `serializationOptions` {@link SerializationOptions} object.
+ *
+ * This is a legacy version of {@link serializerWith}, kept for backward compatibility, and to be deprecated and removed later.
  */
-export const nodeSerializer = <NT extends Node>(reader: Reader<NT>, serializationOptions?: SerializationOptions) => {
+export const nodeSerializer = <NT extends Node>(reader: Reader<NT>, serializationOptions?: SerializationOptions): Serializer<NT> =>
+    serializerWith({ reader, ...serializationOptions })
+
+
+/**
+ * @return a {@link Serializer} function that serializes the {@link Node nodes} passed to it,
+ * configured through a `configuration` {@link SerializerConfiguration} object.
+ */
+export const serializerWith = <NT extends Node>(configuration: SerializerConfiguration<NT>): Serializer<NT> => {
+    const { reader } = configuration
     const propertyValueSerializer =
-        serializationOptions?.propertyValueSerializer ?? serializationOptions?.primitiveTypeSerializer ?? builtinPropertyValueSerializer
-    const serializeEmptyFeatures = serializationOptions?.serializeEmptyFeatures ?? true
+        configuration.propertyValueSerializer ?? configuration.primitiveTypeSerializer ?? builtinPropertyValueSerializer
+    const serializeEmptyFeatures = configuration.serializeEmptyFeatures ?? true
 
     const languageKey2version2classifierKey2allFeatures: Nested3Map<Feature[]> = {}
     const memoisedAllFeaturesOf = (classifier: Classifier): Feature[] =>