+++ packages/test/src/core/library.test.ts
@@ -5,24 +5,25 @@ import {
     dynamicWriter,
     lioncoreBuiltins,
     nameBasedClassifierDeducerFor,
-    nodeSerializer,
-    serializeLanguages
+    serializeLanguages,
+    serializerWith
 } from "@lionweb/core"
 
 import { libraryModel, libraryReader, libraryWriter } from "../instances/library.js"
 import { libraryLanguage } from "../languages/library.js"
 import { deepEqual } from "../test-utils/assertions.js"
 
+
 describe("Library test model", () => {
     it("[de-]serialize example library", () => {
-        const serializationChunk = nodeSerializer(libraryReader)(libraryModel)
+        const serializationChunk = serializerWith({ reader: libraryReader })(libraryModel)
         // FIXME  ensure that serialization does not produce key-value pairs with value === undefined
         const deserialization = deserializeSerializationChunk(serializationChunk, libraryWriter, [libraryLanguage], [])
         deepEqual(deserialization, libraryModel)
     })
 
     it(`"dynamify" example library through serialization and deserialization using the DynamicNode facades`, () => {
-        const serializationChunk = nodeSerializer(libraryReader)(libraryModel)
+        const serializationChunk = serializerWith({ reader: libraryReader })(libraryModel)
         const dynamification = deserializeSerializationChunk(serializationChunk, dynamicWriter, [libraryLanguage], [])
         deepEqual(dynamification.length, 2)
         const lookup = nameBasedClassifierDeducerFor(libraryLanguage)