+++ packages/test/src/core/multi.test.ts
@@ -1,4 +1,4 @@
-import { deserializeSerializationChunk, nodeSerializer } from "@lionweb/core"
+import { deserializeSerializationChunk, serializerWith } from "@lionweb/core"
 import { libraryWriter } from "../instances/library.js"
 import { multiModel, multiReader } from "../instances/multi.js"
 
@@ -6,9 +6,10 @@ import { libraryLanguage } from "../languages/library.js"
 import { multiLanguage } from "../languages/multi.js"
 import { deepEqual } from "../test-utils/assertions.js"
 
+
 describe("multi-language test model", () => {
     it("[de-]serialize multi-language model", () => {
-        const serializationChunk = nodeSerializer(multiReader)(multiModel)
+        const serializationChunk = serializerWith({ reader: multiReader })(multiModel)
         const deserialization = deserializeSerializationChunk(
             serializationChunk,
             libraryWriter,
@@ -18,3 +19,4 @@ describe("multi-language test model", () => {
         deepEqual(deserialization, multiModel)
     })
 })
+