+++ packages/test/src/utilities/textualizer.test.ts
@@ -1,4 +1,4 @@
-import { nodeSerializer } from "@lionweb/core"
+import { serializerWith } from "@lionweb/core"
 import { genericAsTreeText, languageAsText } from "@lionweb/utilities"
 import { readFileSync, writeFileSync } from "fs"
 
@@ -6,6 +6,7 @@ import { libraryModel, libraryReader } from "../instances/library.js"
 import { libraryLanguage } from "../languages/library.js"
 import { languageWithEnum } from "../languages/with-enum.js"
 import { equal } from "../test-utils/assertions.js"
+
 
 describe("LionCore-specific textual syntax", () => {
     it("textualize language with an enum as text", () => {
@@ -19,7 +20,7 @@ describe("LionCore-specific textual syntax", () => {
 describe("generic textual syntax", () => {
     it("textualize library model without language def.", () => {
         equal(
-            genericAsTreeText(nodeSerializer(libraryReader)(libraryModel)),
+            genericAsTreeText(serializerWith({ reader: libraryReader })(libraryModel)),
             `[Library] (id: jkxERSov0TuSh7MMz2D5ciLrZDrU4o_VMOpmBqh_j7E) {
     [library_Library_name] = Bob's Library
     [books]:
@@ -40,7 +41,7 @@ describe("generic textual syntax", () => {
 
     it("textualize library model with language def.", () => {
         equal(
-            genericAsTreeText(nodeSerializer(libraryReader)(libraryModel), [libraryLanguage]),
+            genericAsTreeText(serializerWith({ reader: libraryReader })(libraryModel), [libraryLanguage]),
             `Library (id: jkxERSov0TuSh7MMz2D5ciLrZDrU4o_VMOpmBqh_j7E) {
     name = 'Bob's Library'
     books: