+++ packages/core/CHANGELOG.md
@@ -5,6 +5,8 @@
 * The `NT` type parameter of `SingleRef<NT>` now extends the `Node` base type.
 * Move `IdOrUnresolved` type to the `class-core` package, as it was used only there and in `delta-protocol-common`.
     * Remove the `serializedRef` function as it was a duplicate of `idFrom` in `class-core`.
+* To be able to introduce multiple LionWeb versions explicitly:
+    * Introduce `Serializer` and `SerializerConfiguration` types, and `serializerWith` function as a replacement (keeping an alias for backward compatibility) for the `nodeSerializer` function with a better API (using the Parameter Object pattern).
 
 
 ## 0.7.2